feat(StakeManager): allow for setting rewards when pending period
During testing, we've run into the scenario where we set some rewards for a certain period, but realized that the amount that was set was incorrect.

Updating the rewards is not possible when a reward period is pending. This commit changes that, such that it adds the new rewards to the remaining ones and extends the reward period accordingly.