refactor(Karma)!: `balanceOfRewardDistributor()` -> `actualBalanceOf()`
This commit changes `balanceOfRewardDistributor()` to `actualBalanceOf()` and removes the requirement
of the passed account to be a reward distributor, such that it can be called with any account.

The reason this is done is because we noticed a vulnerability in the voting functionality where
users can cause double accounting by delegating virtual Karma to themselves before redeeming.

The solution will be to have the voting capabilities only use that *actual* Karma for voting
and exclude virtual Karma.