fix(StakeManager)!: don't allow registration of vaults with incorrect owners
This commit fixes a bug that anyone can create `StakeVault`'s for anyone,
which enables DDOSing other users.

The fix is allowing registration of `StakeVault`'s in `StakeManager.registerVault()`
only from a trusted address/account, which in our case is going to be the `VaultFactory`.

Previously, we required that the sender needs a trusted codehash (which is any `StakeVault`
source that we whitelist). While this ensures only correct `StakeVault`s are registered with
the system, it doesn't prevent other users from creating vaults for other accounts, as
setting the `owner` of a `StakeVault` is configurable in its initialization.

The `VaultFactory` ensures that the account that creates the vault is also the owner of the vault.

BREAKING CHANGES:

- `StakeVault.register()` has been removed. Registration is now entirely done via `VaultFactory`.
- `StakeManager.registerVault()` now takes the address of the vault to register as argument:

```diff
- stakeManager.registerVault();
+ stakeManager.registerVault(stakeVaultAddress);
```

Addresses https://github.com/Cyfrin/audit-2025-12-statusl2/issues/1

Closes #72