fix(StakeManager): prevent migrating to non-empty vaults
The check in `migrateToVault()` that ensures a vault has `0` staked balanace is not enough to ensure the vault we're
migrating to is in fact empty. Users could've unstaked in the past. This is by itself not an issue, but if the vault has
already accumulated rewards, then those will still live in `rewardsAccrued`.

The migration will override the this value with whatever rewards have been accrued in the vault the user is migrating
from. This can erase previously accrued rewards in the vault the user is migrating to.

This commit fixes this by extending the check on `rewardsAccrued`.

Addresses https://github.com/Cyfrin/audit-2025-12-statusl2/issues/29